backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
unstructured[md,pdf,docx]==0.10.30
pymupdf==1.23.5
celery==5.3.4
redis==5.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1 